% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExchangeRates.R
\name{convert_exchange_rates}
\alias{convert_exchange_rates}
\title{Convert a value to a specific range of other currencies from the Currencyapi.com API. Retrieve the latest exchange rate data or for a specific historic date.}
\usage{
convert_exchange_rates(
  value,
  date = NULL,
  base_currency = NULL,
  currencies = NULL
)
}
\arguments{
\item{value}{(required) The value you want to convert}

\item{date}{Date to retrieve historical rates from (format: 2021-12-31)}

\item{base_currency}{The base currency for the conversion (e.g., "USD")}

\item{currencies}{A list of comma seperated currency codes which you want to get (EUR,USD,CAD) By default all available currencies will be shown}
}
\value{
Returns calculated values for today or any given date for all or a specific set of currencies.
}
\description{
Convert a value to a specific range of other currencies from the Currencyapi.com API. Retrieve the latest exchange rate data or for a specific historic date.
}
