% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateRangeHistoricalExchangeRates.R
\name{get_date_range_historical_exchange_rates}
\alias{get_date_range_historical_exchange_rates}
\title{Get historical exchange rates for a specific date for a single or multiple currencies.}
\usage{
get_date_range_historical_exchange_rates(
  datetime_start,
  datetime_end,
  accuracy = NULL,
  base_currency = NULL,
  currencies = NULL
)
}
\arguments{
\item{datetime_start}{(required) Datetime for the start of your requested range (format: 2021-12-31T23:59:59Z / ISO8601 Datetime)}

\item{datetime_end}{(required) Datetime for the end of your requested range (format: 2021-12-31T23:59:59Z / ISO8601 Datetime)}

\item{accuracy}{The accuracy you want to receive; Possible Values: day, hour, quarter_hour, minute; Default: day For valid time ranges see below}

\item{base_currency}{The base currency for the conversion (e.g., "USD")}

\item{currencies}{A list of comma seperated currency codes which you want to get (EUR,USD,CAD) By default all available currencies will be shown}
}
\value{
Returns exchange rates for a given time range. Generally, we provide data going back to 1999.
}
\description{
Get historical exchange rates for a specific date for a single or multiple currencies.
}
